+++ src/hooks/useRecipeFiltering.ts
@@ -1,0 +1,50 @@
+import { useState, useEffect, useMemo } from 'react';
+import { useParams, useNavigate } from 'react-router-dom';
+import { Recipe } from '../types/Recipe';
+
+/**
+ * Custom hook for recipe filtering and category management
+ * Handles URL synchronization, category selection, and recipe filtering
+ */
+export const useRecipeFiltering = (recipes: Recipe[]) => {
+  const { category: urlCategory } = useParams<{ category: string }>();
+  const navigate = useNavigate();
+  const [selectedCategory, setSelectedCategory] = useState<string | null>(null);
+
+  // Get unique categories from all recipes
+  const categories = useMemo(() => 
+    Array.from(new Set(recipes.map(recipe => recipe.category))),
+    [recipes]
+  );
+
+  // Sync URL category with state
+  useEffect(() => {
+    setSelectedCategory(urlCategory || null);
+  }, [urlCategory]);
+
+  // Toggle category selection and update URL
+  const toggleCategory = (category: string) => {
+    if (selectedCategory === category) {
+      // Deselect - go back to home
+      navigate('/');
+    } else {
+      // Select - navigate to category URL
+      navigate(`/category/${category}`);
+    }
+  };
+
+  // Filter recipes based on selected category
+  const filteredRecipes = useMemo(() =>
+    selectedCategory
+      ? recipes.filter(recipe => recipe.category === selectedCategory)
+      : recipes,
+    [recipes, selectedCategory]
+  );
+
+  return {
+    selectedCategory,
+    categories,
+    filteredRecipes,
+    toggleCategory
+  };
+};
